+++ Cargo.toml
@@ -91,6 +91,8 @@ ssi-jwk = { path = "./ssi-jwk" }
 ssi-json-ld = { path = "./ssi-json-ld" }
 ssi-core = { path = "./ssi-core" }
 ssi-dids = { path = "./ssi-dids" }
+ssi-crypto = { path = "./ssi-crypto" }
+caips = { path = "./caips" }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 clear_on_drop = "0.2.4"
