+++ src/ssh.rs
@@ -1,4 +1,4 @@
-use jwk::{Base64urlUInt, Params as JWKParams, JWK};
+use crate::jwk::{Base64urlUInt, Params as JWKParams, JWK};
 use sshkeys::PublicKeyKind;
 use thiserror::Error;
 
@@ -17,7 +17,7 @@ pub enum SSHKeyToJWKError {
 }
 
 fn pk_to_jwk_rsa(pk: &sshkeys::RsaPublicKey) -> JWK {
-    JWK::from(JWKParams::RSA(jwk::RSAParams::new_public(
+    JWK::from(JWKParams::RSA(crate::jwk::RSAParams::new_public(
         &pk.e, &pk.n,
     )))
 }
@@ -31,8 +31,7 @@ fn pk_to_jwk_ecdsa(pk: &sshkeys::EcdsaPublicKey) -> Result<JWK, SSHKeyToJWKError
             }
             #[cfg(feature = "p256")]
             {
-                jwk::p256_parse(&pk.key)
-                    .map_err(|e| SSHKeyToJWKError::P256Parse(e.to_string()))
+                jwk::p256_parse(&pk.key).map_err(|e| SSHKeyToJWKError::P256Parse(e.to_string()))
             }
         }
         _ => Err(SSHKeyToJWKError::UnsupportedEcdsaKey(
@@ -42,7 +41,7 @@ fn pk_to_jwk_ecdsa(pk: &sshkeys::EcdsaPublicKey) -> Result<JWK, SSHKeyToJWKError
 }
 
 fn pk_to_jwk_ed25519(pk: &sshkeys::Ed25519PublicKey) -> JWK {
-    JWK::from(JWKParams::OKP(jwk::OctetParams {
+    JWK::from(JWKParams::OKP(crate::jwk::OctetParams {
         curve: "Ed25519".to_string(),
         public_key: Base64urlUInt(pk.key.clone()),
         private_key: None,
